# swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "This is a simple example NodeJS API project to demonstrate Swagger Documentation",
      "version": "1.0.0",
      "title": "Tasks API",
      "contact": {
        "email": "abc@gmail.com"
      },
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      }
    },
    "schemes": [
      "http"
    ],
    "host": "localhost:4501",
    "basePath": "/api",
    "paths": {
      "/tasks": {
        "get": {
          "summary": "Get all the tasks",
          "description": "Get all the tasks",
          "produces": [
            "application/json"
          ],
          "parameters": [
            
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/todosResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      },
      "/task": {
        "post": {
          "summary": "Save the task",
          "description": "Save the task",
          "produces": [
            "application/json"
          ],
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "task object",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "type": "object",
                    "$ref": "#/definitions/Task"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/todosResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      },
      "/task/": {
        "put": {
          "summary": "Update the tasks",
          "description": "Update the tasks",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "task id that needs to be deleted",
              "required": true,
              "type": "string"
            },
            {
              "in": "body",
              "name": "body",
              "description": "task object",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "type": "object",
                    "$ref": "#/definitions/Task"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/todosResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      },
      "/task/{id}": {
        "delete": {
          "summary": "Delete the task",
          "description": "Delete the task",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "task id that needs to be deleted",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/todosResponse"
                }
              }
            },
            "400": {
              "description": "Invalid status value",
              "schema": {
                "$ref": "#/definitions/InvalidResponse"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "todosResponse": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "task": {
            "type": {
              "name": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "createdBy": {
                "type": "string"
              }
            }
          }
        }
      },
      "InvalidResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }